clean up clientnum code

+++ include/cmds.h
@@ -17,7 +17,7 @@ Created By:
 
 #include "main.h"
 
-typedef int (*pfnAdminCmd)(int clientnum, int access, std::vector<std::string> args, bool say);		// signature of a command handler
+typedef int (*pfnAdminCmd)(intptr_t clientnum, int access, std::vector<std::string> args, bool say);		// signature of a command handler
 
 // command handler info
 typedef struct admincmd_s {
@@ -33,7 +33,7 @@ extern std::vector<admincmd_t> g_admincmds;
 extern std::vector<admincmd_t> g_saycmds;
 
 void reload();
-int handlecommand(int clientnum, std::vector<std::string> args);
+int handlecommand(intptr_t clientnum, std::vector<std::string> args);
 int admin_adduser(addusertype_t type, std::vector<std::string> args);
 
 #endif // __QADMIN_QMM_UTIL_H__
